pyproject.toml
[tool.poetry]
name = "python-thermal-printer"
version = "0.1.0"
description = ""
authors = ["Robert DeLanghe <bdelanghe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
Pillow = "^7.1.1"
tomlkit = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "<2.0.0"
black = "^19.10b0"
mypy = "^0.770"
tox = "^3.14.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"